Refactorings on DatagramEndpoint
- Added Interface for DatagramEndpoint
- Better encapsulation for packet
- Added JavaDoc
