+++ client/src/main/java/ch/bifrost/client/test/UDPTestClient.java
@@ -1,7 +1,7 @@
 package ch.bifrost.client.test;
 
-import ch.bifrost.core.api.datagram.Packet;
-import ch.bifrost.core.api.datagram.UDPDatagramEndpoint;
+import ch.bifrost.core.impl.datagram.Packet;
+import ch.bifrost.core.impl.datagram.UDPDatagramEndpoint;
 
 public class UDPTestClient {
