--- core/src/main/java/ch/bifrost/core/Hello.java
@@ -1,7 +1,0 @@
-package ch.bifrost.core;
-
-public class Hello {
-	public static void main (String[] args) {
-		System.out.println("hello there");
-	}
-}
