+++ core/src/main/java/ch/bifrost/core/impl/datagram/UDPDatagramEndpoint.java
@@ -1,6 +1,5 @@
-package ch.bifrost.core.api.datagram;
+package ch.bifrost.core.impl.datagram;
 
-import java.io.Closeable;
 import java.io.IOException;
 import java.net.DatagramPacket;
 import java.net.DatagramSocket;
@@ -8,18 +7,32 @@ import java.net.SocketException;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 
-public class UDPDatagramEndpoint implements Closeable {
+import ch.bifrost.core.api.datagram.DatagramEndpoint;
+
+/**
+ * Implementation of {@link DatagramEndpoint} for UDP.
+ */
+public class UDPDatagramEndpoint implements DatagramEndpoint {
 
 	private static final int SOCKET_TIMEOUT_IN_MS = 1000;
 	private DatagramSocket socket;
 	private BlockingQueue<Packet> receivedPackets = new LinkedBlockingQueue<>();
 	private Receiver receiver;
 
+	/**
+	 * Create an endpoint that listens on a specific port (usually required for the server). 
+	 * @param port the port to listen on
+	 * @throws SocketException when the endpoint cannot be created
+	 */
 	public UDPDatagramEndpoint(int port) throws SocketException {
 		socket = new DatagramSocket(port);
 		initialize();
 	}
 
+	/**
+	 * Create an endpoint on any arbitrary port.
+	 * @throws SocketException thrown if endpoint cannot be created
+	 */
 	public UDPDatagramEndpoint() throws SocketException {
 		socket = new DatagramSocket();
 		initialize();
@@ -42,11 +55,12 @@ public class UDPDatagramEndpoint implements Closeable {
 		socket.close();
 	}
 	
+	@Override
 	public void send(Packet outgoing) throws IOException {
-		DatagramPacket datagram = new DatagramPacket(outgoing.getBytes(), outgoing.getLength(), outgoing.getCounterpartAddress(), outgoing.getCounterpartPort());
-		socket.send(datagram);
+		socket.send(outgoing.toDatagram());
 	}
 	
+	@Override
 	public Packet receive() throws InterruptedException {
 		return receivedPackets.take();
 	}
@@ -72,7 +86,7 @@ public class UDPDatagramEndpoint implements Closeable {
 				} catch (IOException e) {
 					continue;
 				}
-				Packet packet = new Packet(datagram.getAddress(), datagram.getPort(), datagram.getData(), datagram.getLength());
+				Packet packet = Packet.fromDatagram(datagram);
 				try {
 					queue.put(packet);
 				} catch (InterruptedException e) {
